fix: address all nateinaction review comments
Agent-Logs-Url: https://github.com/Open-Source-Space-Foundation/proves-core-reference/sessions/bef208f4-9553-431a-9360-8b48d7da9d42
