+++ Makefile
@@ -227,7 +227,7 @@ test-unit: ## Run unit tests
 test-integration: uv ## Run integration tests (set TEST=<name|file.py> or pass test targets)
 	@DEPLOY="build-artifacts/zephyr/fprime-zephyr-deployment"; \
 	TARGETS=""; \
-	FILTER="not flaky and not requires_hw_watchdog"; \
+	FILTER="not flaky"; \
 	if [ -n "$(TEST)" ]; then \
 		case "$(TEST)" in \
 			*.py) TARGETS="PROVESFlightControllerReference/test/int/$(TEST)" ;; \
@@ -250,13 +250,13 @@ test-integration: uv ## Run integration tests (set TEST=<name|file.py> or pass t
 
 .PHONY: reenable-lora
 reenable-lora: uv ## Re-enable LoRa transmit before the RF integration pass
-	$(UV_RUN) pytest PROVESFlightControllerReference/test/int/0_radio_test.py::test_00_setup_only \
+	$(UV_RUN) pytest PROVESFlightControllerReference/test/int/radio_test.py::test_00_setup_only \
 	    --deployment build-artifacts/zephyr/fprime-zephyr-deployment
 
 .PHONY: test-integration-radio
 test-integration-radio: uv ## Run integration tests via LoRa passthrough GDS
 	@DEPLOY="build-artifacts/zephyr/fprime-zephyr-deployment"; \
-	FILTER="not flaky and not requires_hw_watchdog and not rf_unsafe"; \
+	FILTER="not flaky and not rf_unsafe"; \
 	$(UV_RUN) pytest PROVESFlightControllerReference/test/int \
 	    --deployment $$DEPLOY -m "$$FILTER"
