+++ pytest.ini
@@ -2,4 +2,3 @@
 markers =
     slow: marks tests as slow
 	flaky: marks tests as flaky
-	requires_hw_watchdog: needs HW watchdog wired/enabled (skipped on v5e for now)
